agents/devolucoes/pyproject.toml
[project]
name = "agent-devolucoes"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = ["mcp-core"]

[project.scripts]
agent-server = "agent.server:main"

[tool.uv.sources]
mcp-core = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]

[tool.hatch.build.targets.wheel]
packages = ["src/agent"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: mark test as an integration test (requires running server)",
]
